Create twopointer
onverts the input string into a character array using the toCharArray() method of the String class. It then initializes two pointers, i and j, which point to the first and last characters of the array, respectively. It then iterates through the array using the while loop, swapping the characters pointed to by i and j on each iteration until the pointers meet in the middle. Finally, it returns the reversed string by creating a new String object from the reversed character array using the String(char[]) constructor.